Moved the donation form to its own component
