+++ components/DonationForm/index.tsx
@@ -1,0 +1,3 @@
+import DonationForm from './DonationForm';
+
+export default DonationForm;
